+++ packages/icons/src/components/action-search/action-search.tsx
@@ -4,11 +4,20 @@ import { Component, Prop, Host, h } from '@stencil/core';
   tag: 'scale-icon-action-search',
 })
 export class ActionSearch {
-  @Prop() size: number = 24;
-  @Prop() color: string = 'currentColor';
-  @Prop() selected: boolean = false;
+  /** (optional) The width and height in pixels */
+  @Prop() size?: number = 24;
+  /** (optional) Sets the icon color via the `fill` attribute */
+  @Prop() color?: string = 'currentColor';
+  /** (optional) If `true`, the icon changes visually */
+  @Prop() selected?: boolean = false;
+  /** (optional) If `true` the SVG element will get `aria-hidden="true"` */
+  @Prop() decorative?: boolean = false;
+  /** (optional) When using the icon standalone, make it meaningful for accessibility */
+  @Prop() accessibilityTitle?: string;
 
   render() {
+    const ariaHidden = this.decorative ? { 'aria-hidden': 'true' } : {};
+
     return (
       <Host style={{ display: 'inline-flex' }}>
         <svg
@@ -16,7 +25,9 @@ export class ActionSearch {
           width={this.size}
           height={this.size}
           viewBox="0 0 24 24"
+          {...ariaHidden}
         >
+          {this.accessibilityTitle && <title>{this.accessibilityTitle}</title>}
           <g fill={this.color}>
             {this.selected ? (
               <g>