+++ packages/storybook-vue/stories/3_components/icon-library/ExampleIcon.vue
@@ -1,6 +1,12 @@
 <template>
   <div>
-    <scale-icon-action-favorite :color="color" :size="size" :selected="selected" />
+    <scale-icon-action-favorite
+      :color="color"
+      :size="size"
+      :selected="selected"
+      :decorative="decorative"
+      :accessibility-title="accessibilityTitle"
+    />
   </div>
 </template>
 
@@ -9,7 +15,9 @@ export default {
   props: {
     color: { type: String, default: 'currentColor' },
     size: { type: Number, default: 24 },
-    selected: { type: Boolean, default: false }
+    selected: { type: Boolean, default: false },
+    decorative: { type: Boolean, default: false },
+    accessibilityTitle: { type: String },
   }
 }
 </script>